refactor: split large files with proper access control
- Split SessionManager+SplitPane.swift (fileprivate → internal access)
- Split SOCKS5Handler.swift (add Citadel import)
- Split PaneTerminalView+ContextMenu.swift (internal access)
- Extract LocalFileRow.swift from SFTPWindowView.swift
- Fix NavigationDirection enum naming (.up → .upward, .down → .downward)
- Fix SFTPService.swift explicit self references
- Fix SerialPortService.swift Double/Int type mismatch
- Fix WindowLayout.swift variable naming (i → index)
- Zero SwiftLint violations, zero SwiftFormat errors

+++ Bonk/Services/SerialPort/SerialPortService.swift
@@ -170,7 +170,7 @@ final class SerialPortService {
         }
     }
 
-    private func setStopBits(_ bits: Int, tty: inout termios) {
+    private func setStopBits(_ bits: Double, tty: inout termios) {
         if bits == 2 {
             tty.c_cflag |= UInt(CSTOPB)
         } else {
